requirements.txt
jax
optax
dm-haiku
einops
